pyproject.toml
[tool.poetry]
name = "moz-books"
version = "0.1.0"
description = "mozkzki's book service library."
authors = ["mozkzki <mozkzki@gmail.com>"]
maintainers = ["mozkzki <mozkzki@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/mozkzki/moz-books"
repository = "https://github.com/mozkzki/moz-books"
documentation = "https://github.com/mozkzki/moz-books/wiki"
keywords = ['book', 'rakuten', 'calil']
classifiers=[
    'Development Status :: 3 - Alpha',
    'License :: OSI Approved :: MIT License',
    'Topic :: Software Development :: Libraries',
    'Intended Audience :: Developers',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.19.0"
requests = "2.25.1"
coloredlogs = "^15.0.1"

[tool.poetry.dev-dependencies]
flake8 = "3.9.2"
mypy = "0.910"
pytest = "6.2.5"
black = "21.9b0"
pytest-cov = "2.12.1"
coverage = "5.5"
isort = "5.9.3"
Sphinx = "^4.2.0"
sphinx-rtd-theme = "^1.0.0"
types-requests = "^2.25.9"
pytest-env = "^0.6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
